Here is a massive patch that does nearly nothing.
It adds the type definition for a structure called  
pdf_output_file. Currently it contains only the FILE * 
of the output file, but soon this will be filled in
with more of the currently still global variables
related to pdf writing.

The patch also adds a pointer to such a structure to 
all the functions that write to the PDF output (or so I hope).

There is a new global variable 'static_pdf' (in pdfgen.c),
which contains the reference to the sole object of the 
pdf_output_file type. There is still a need for such
a static variable because there are hard to reach pdf 
writing functions (esp. in limglib.c, lpdflib.c, and 
pdftoepdf.cc's read_pdf_info() function).

The suggested entry point is the check_pdfminorversion() 
function, as that creates the global structure if it does 
not exist yet.

The only other function that explicitly test the variable
for existance is remove_pdffile(), everything else just
assumes a non-NULL argument.

Oh yeah: this patch also shifts a few minor things
around, like moving do_late_lua() to pdfshipout.c.



git-svn-id: http://foundry.supelec.fr/svn/luatex/trunk@2579 0b2b3880-5936-4365-a048-eb17d2e5a6bf